Gas Resistance Profile + OLED Display
The first part of the the code above is the inserting of Gas Resistance (from the BME680 sensor) values into a total of 10 "step" values. In other words, the sensor reads the Gas Resistance 10 times and inserts it into 10 different "step" values. 

The second part of the code is the OLED display. If, else if, and else statements are used to determine whether each individual Gas Resistance value is above 60 KOhms or not. If one of the values is below 60 KOhms, the OLED will display "Caution: VOC Detected". Otherwise, the OLED would display "No VOC Detected". Since this part of the code is in the void loop() section, it continues to run and will automatically change between "VOC detected" and "no VOC detected" when needed.