Visual Studio Data Plotting Code 8 - Python
This code is in the language of python and is used in Visual Studio for the VOC project. Additional details are in the README section. The code focuses on gathering data from the BME680 sensor and examples/tests in Arduino IDE, sorting the data, and plotting the data on a bar graph. This helps visualize the changes and variation of Gas Resistance (or other values) over time.