BME680 Test
This file links to an Arduino IDE test that focuses on printing temperature, humidity, pressure, and gas. This code file/test linked comments out data headers or titles for the values in order to use the Plot Data 7 Visual Studio Code (linked in the repository). These comments are based on preference and can be taken out (they start with Serial.print() and are easy to find). I recommend taking out the comments beginners to understand what each value of code is for. 

The code will continue to loop and print temperature, humidity, pressure, and gas values each time.