Visual Studio Data Plotting Code 7 - Python
This Visual Studio Python Data Plotting Code is the same as the Visual Studio Data Plotting Code 8 file; however, this file was modified to separate the data list for the BME680 test (which is also a temperature reading file) in order to only plot data for the temperature. I have also modified the code to print in Fahrenheit, but if preferred in Celsius, the  farTemp line in the while loop can be adjusted and commented out.